refactor: ForecastResultItem 재구성 및 dto 클래스 구조 재구성
