+++ src/main/java/com/backendoori/ootw/weather/dto/forecast/BaseDateTime.java
@@ -1,4 +1,4 @@
-package com.backendoori.ootw.weather.domain.forecast;
+package com.backendoori.ootw.weather.dto.forecast;
 
 public record BaseDateTime(
     String baseDate,
