+++ src/main/java/com/backendoori/ootw/weather/service/WeatherService.java
@@ -3,9 +3,9 @@ package com.backendoori.ootw.weather.service;
 import java.time.LocalDateTime;
 import java.util.HashMap;
 import java.util.Map;
-import com.backendoori.ootw.weather.domain.forecast.BaseDateTime;
 import com.backendoori.ootw.weather.domain.forecast.ForecastCategory;
 import com.backendoori.ootw.weather.dto.WeatherResponse;
+import com.backendoori.ootw.weather.dto.forecast.BaseDateTime;
 import com.backendoori.ootw.weather.util.BaseDateTimeCalculator;
 import com.backendoori.ootw.weather.util.client.ForecastApiClient;
 import lombok.RequiredArgsConstructor;
