+++ src/main/java/com/backendoori/ootw/weather/util/deserializer/ForecastResultHeaderDeserializer.java
@@ -1,22 +1,18 @@
 package com.backendoori.ootw.weather.util.deserializer;
 
 import java.io.IOException;
-import com.backendoori.ootw.weather.domain.forecast.ForecastResultHeader;
+import com.backendoori.ootw.weather.dto.forecast.ForecastResultHeader;
 import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.databind.DeserializationContext;
 import com.fasterxml.jackson.databind.JsonDeserializer;
 import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
 public class ForecastResultHeaderDeserializer extends JsonDeserializer<ForecastResultHeader> {
 
-    private final ObjectMapper objectMapper;
-
     @Override
-    public ForecastResultHeader deserialize(JsonParser p, DeserializationContext ctxt)
-        throws IOException {
+    public ForecastResultHeader deserialize(JsonParser p, DeserializationContext ctxt) throws IOException {
         JsonNode node = p.getCodec()
             .readTree(p);
         JsonNode headerNode = node.findValue("response")
