+++ src/main/java/com/backendoori/ootw/weather/util/deserializer/ForecastSuccessResultBodyDeserializer.java
@@ -4,8 +4,8 @@ import java.io.IOException;
 import java.util.Arrays;
 import java.util.List;
 import com.backendoori.ootw.weather.domain.forecast.ForecastCategory;
-import com.backendoori.ootw.weather.domain.forecast.ForecastResultItem;
-import com.backendoori.ootw.weather.domain.forecast.ForecastSuccessResultBody;
+import com.backendoori.ootw.weather.dto.forecast.ForecastResultItem;
+import com.backendoori.ootw.weather.dto.forecast.ForecastSuccessResultBody;
 import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.databind.DeserializationContext;
 import com.fasterxml.jackson.databind.JsonDeserializer;
